+++ client/src/api/msg.c
@@ -1,6 +1,16 @@
 #include "client.h"
 
-t_dtp *mx_msg_request(char *msg, int room_id) {
+/*
+ * Function: mx_msg_request
+ * -------------------------------
+ * Creates message with TEXT type request
+ * 
+ * msg: message
+ * room_id: room id for message
+ * 
+ * returns: message request
+ */
+t_dtp *mx_msg_request(char *msg, guint64 room_id) {
     cJSON *json_result = cJSON_CreateObject();
 
     if (!cJSON_AddNumberToObject(json_result, "type", RQ_MSG))
@@ -14,6 +24,16 @@ t_dtp *mx_msg_request(char *msg, int room_id) {
     return mx_get_transport_data(json_result);
 }
 
+/*
+ * Function: mx_msg_handler
+ * -------------------------------
+ * Handles request from server
+ * 
+ * data: request from server
+ * chat: information about chat
+ * 
+ * returns: success of handling
+ */
 gboolean mx_msg_handler(t_dtp *data, t_chat *chat) {
     t_gmsg *gmsg = mx_create_gmsg(data->json, chat);
     t_groom *groom = NULL;
